update most stories
