+++ packages/storybook-react/stories/extension-hard-break/hard-break.stories.tsx
@@ -1,36 +1,3 @@
-import 'remirror/styles/all.css';
-
-import { htmlToProsemirrorNode } from 'remirror';
-import { HardBreakExtension } from 'remirror/extensions';
-import { Remirror, ThemeProvider, useCommands, useRemirror } from '@remirror/react';
-
+import Basic from './basic';
+export { Basic };
 export default { title: 'Extensions / HardBreak' };
-
-const extensions = () => [new HardBreakExtension()];
-
-const HardBreakButton = () => {
-  const commands = useCommands();
-  return <button onClick={() => commands.insertHardBreak()}>Insert</button>;
-};
-
-export const Basic = (): JSX.Element => {
-  const { manager, state, onChange } = useRemirror({
-    extensions: extensions,
-    content: '<p>Text with <br />hard break</p>',
-    stringHandler: htmlToProsemirrorNode,
-  });
-
-  return (
-    <ThemeProvider>
-      <Remirror
-        manager={manager}
-        autoFocus
-        onChange={onChange}
-        initialContent={state}
-        autoRender='end'
-      >
-        <HardBreakButton />
-      </Remirror>
-    </ThemeProvider>
-  );
-};
