+++ packages/storybook-react/stories/extension-image/basic.tsx
@@ -1,0 +1,31 @@
+import 'remirror/styles/all.css';
+
+import React from 'react';
+import { htmlToProsemirrorNode } from 'remirror';
+import { ImageExtension } from 'remirror/extensions';
+import { Remirror, ThemeProvider, useRemirror } from '@remirror/react';
+
+const extensions = () => [new ImageExtension()];
+
+const Basic: React.FC = () => {
+  const { manager, state, onChange } = useRemirror({
+    extensions,
+    content:
+      '<p>You can see a green image below.</p><img src="https://dummyimage.com/200x80/479e0c/fafafa">',
+    stringHandler: htmlToProsemirrorNode,
+  });
+
+  return (
+    <ThemeProvider>
+      <Remirror
+        manager={manager}
+        autoFocus
+        onChange={onChange}
+        initialContent={state}
+        autoRender='end'
+      />
+    </ThemeProvider>
+  );
+};
+
+export default Basic;
