+++ packages/storybook-react/stories/extension-strike/basic.tsx
@@ -1,0 +1,42 @@
+import 'remirror/styles/all.css';
+import './styles.css';
+
+import { cx, htmlToProsemirrorNode } from 'remirror';
+import { StrikeExtension } from 'remirror/extensions';
+import { Remirror, ThemeProvider, useActive, useCommands, useRemirror } from '@remirror/react';
+
+const extensions = () => [new StrikeExtension()];
+
+const StrikeButton = () => {
+  const commands = useCommands();
+  const active = useActive(true);
+  return (
+    <button onClick={() => commands.toggleStrike()} className={cx(active.strike() && 'active')}>
+      Strike
+    </button>
+  );
+};
+
+const Basic = (): JSX.Element => {
+  const { manager, state, onChange } = useRemirror({
+    extensions: extensions,
+    content: '<p>Text in <strike>strike</strike></p>',
+    stringHandler: htmlToProsemirrorNode,
+  });
+
+  return (
+    <ThemeProvider>
+      <Remirror
+        manager={manager}
+        autoFocus
+        onChange={onChange}
+        initialContent={state}
+        autoRender='end'
+      >
+        <StrikeButton />
+      </Remirror>
+    </ThemeProvider>
+  );
+};
+
+export default Basic;
