+++ packages/storybook-react/stories/extension-underline/basic.tsx
@@ -1,0 +1,45 @@
+import 'remirror/styles/all.css';
+import './styles.css';
+
+import { cx, htmlToProsemirrorNode } from 'remirror';
+import { UnderlineExtension } from 'remirror/extensions';
+import { Remirror, ThemeProvider, useActive, useCommands, useRemirror } from '@remirror/react';
+
+const extensions = () => [new UnderlineExtension()];
+
+const UnderlineButton = () => {
+  const commands = useCommands();
+  const active = useActive(true);
+  return (
+    <button
+      onClick={() => commands.toggleUnderline()}
+      className={cx(active.underline() && 'active')}
+    >
+      Underline
+    </button>
+  );
+};
+
+const Basic = (): JSX.Element => {
+  const { manager, state, onChange } = useRemirror({
+    extensions: extensions,
+    content: '<p>Text in <u>underline</u></p>',
+    stringHandler: htmlToProsemirrorNode,
+  });
+
+  return (
+    <ThemeProvider>
+      <Remirror
+        manager={manager}
+        autoFocus
+        onChange={onChange}
+        initialContent={state}
+        autoRender='end'
+      >
+        <UnderlineButton />
+      </Remirror>
+    </ThemeProvider>
+  );
+};
+
+export default Basic;
